Add Support for Using Custom SSH Key
Despite it being tedious for some users and possibly unnecessary, for extra security and peace of mind, \
I love having all of my blogs and webapps having separate ssh keys for their deployments. Currently, \
adding a custom key through the args parameter doesn't work properly if the port parameter is set \
in the config, but that's an issue for another day and may or may not be a design choice. Enjoy